% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_add_to_database.R
\name{add_to_database}
\alias{add_to_database}
\title{Add a new lake to the database}
\usage{
add_to_database(RSPath, dbName, folder = NULL, df = NULL)
}
\arguments{
\item{RSPath}{Path of remoate sensing input and output folders}

\item{dbName}{Name of the dabase}

\item{folder}{A folder Name of the data to be added.  On of folder or df
needs to be defined}

\item{df}{A dataframe to be added. On of folder or df needs to be defined}
}
\value{
Function saves the extended database table as csv
}
\description{
Add a new lake to the database
}
