% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_NDTrI.R
\name{aggregate_NDTrI}
\alias{aggregate_NDTrI}
\title{Aggregation of all pixel NDTrIs into one lake NDTrI}
\usage{
aggregate_NDTrI(ndtriPixels)
}
\arguments{
\item{ndtriPixels}{A list of NDTrI per pixel of one or more years, created by
\code{\link[=seasonal_index_per_pixel]{seasonal_index_per_pixel()}}}
}
\description{
Aggregation of all pixel NDTrIs into one lake NDTrI
}
