% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_from.R
\name{bbox_from_polygon}
\alias{bbox_from_polygon}
\title{Bounding Box from rectangle}
\usage{
bbox_from_polygon(geom_sfc, meters = 20)
}
\arguments{
\item{geom_sfc}{A simple feature polygon in WGS 84 EPSG:4326.
See example how to create a sfc_POLYGON from Well-known-text (wkt) Geometry.}

\item{meters}{The buffer distance in m around the polygon to create the
bounding box (Needs to be > 0)}
}
\description{
Bounding Box from rectangle
}
\examples{
library(sf)
x <- data.frame(
 "wkt" = paste0(
   "POLYGON ((13.22103 52.61554, 13.22112 52.61558,13.22118 52.6156,",
   " 13.2212 52.61561, 13.22123 52.61561, 13.22127 52.61562))")
)
x$wkt
is(x$wkt)
geom_sfc <- sf::st_as_sf(x = x, wkt = "wkt", crs = 4326)
geom_sfc
is(geom_sfc)

}
