% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_years.R
\name{combine_years_dynamic}
\alias{combine_years_dynamic}
\title{Combine index dynamic of multiple years for a whole lake}
\usage{
combine_years_dynamic(indexDynamicList, years = NULL)
}
\arguments{
\item{indexDynamicList}{A list of index dynamic created by \code{\link[=dynamic_per_pixel]{dynamic_per_pixel()}}}

\item{years}{A numeric vector of all years to be considered. If NULL, all
available years will be used}
}
\description{
Combine index dynamic of multiple years for a whole lake
}
