% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EO_assessment_class.R
\name{determine_classes}
\alias{determine_classes}
\title{The class of NDTrI is determined compared to other lakes per year}
\usage{
determine_classes(yearly_spread, nClass = 10)
}
\arguments{
\item{yearly_spread}{A dataframe of lakeName, lakeID and the yearly NDTrI
values as created by \code{\link[=ndtri_spread]{ndtri_spread()}}}

\item{nClass}{The number of class used (3 classes minimum).}
}
\value{
A list of 4:

"classes": A matrix of classes as characters, where number of rows equal the number of
lakes (in the same order as in the yearly_spread input table) and number of
columns correspond to the number of years assessed
"breaks": The breaks of the classes.
"nClass": the total number of classes
"upperLowerLimit": The proportion of lakes in the first and last class
}
\description{
The class of NDTrI is determined compared to other lakes per year
}
\details{
The first and last class are by default defined as the least and most
eutrophicated 5\% of all lakes, respectively. However, the classes should
always contain less lakes as the average number of lakes per class. The
quantile limit is adjusted if the number of classes is extremely high compared
to the number of lakes. The remaining classes between first and last one,
are equal in range to avoid high sensitivity regions in the ranking,
i.e. the classes can contain different amounts of lakes.
}
