% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericAssessment.R
\name{determine_status}
\alias{determine_status}
\title{The average status of an index over n years}
\usage{
determine_status(yearly_spread, n_years = 3)
}
\arguments{
\item{yearly_spread}{A dataframe of lakes or pixels as row and the yearly
index values as columns as created by \code{\link[=index_spread]{index_spread()}}}

\item{n_years}{The number of years used for the moving average. By defaul
this is 3 years.}
}
\value{
A list of 2:

A matrix, where number of rows equal the number of lakes (in the same order
as in the yearly_spread input table) and number of correspond to the number
of years assessed
}
\description{
The average status of an index over n years
}
