% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericAssessment.R
\name{determine_trend}
\alias{determine_trend}
\title{Trend of an Index over years}
\usage{
determine_trend(yearly_spread = yearly_spread, tyears = 3, ttype = "short")
}
\arguments{
\item{yearly_spread}{A data frame of lakes or pixels as row and the yearly
index values as columns as created by \code{\link[=index_spread]{index_spread()}}}

\item{tyears}{The number of years used for the trend analysis.}

\item{ttype}{The type of trend (either "short" or "long") as character string}
}
\value{
A list of 4:
"trends" is a matrix with one row per lake in the as in the
yearlyspread input table and two columns for each trend, short-term and
long-term. The trend is the slope of the regression line, and error is the
standard error of the slope.
"periods": number of years used for short-term and long-term trend
"overallMedian": median value of the whole dataset
"yearlyMedians": median values of all lakes per year
}
\description{
Trend of an Index over years
}
\details{
In order to rule out systematic errors based on weather conditions in a year,
different hardware or algorithms, the difference between the overall median
value of all lakes (or all pixels) and years and the yearly median value of
all lakes(or all pixels) is added to the Index of a lake (or pixel) before
the trends are calculated. Median values are based on lakes or pixels for which data is available in all considered
years.
Subsequantly, the trend is calculated as linear regression between previous
years and adjusted NDTrI values.
For a reliable trend assessment the number of lakes needs to be large enough
(n>10).
}
