% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_bbox.R
\name{display_bbox}
\alias{display_bbox}
\title{Display the Bounding Box}
\usage{
display_bbox(bbox, zoom = 15)
}
\arguments{
\item{bbox}{The bounding box to disply}

\item{zoom}{The zoom level of the initial map}
}
\description{
Display the Bounding Box
}
