% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_per_pixel.R
\name{dynamic_per_pixel}
\alias{dynamic_per_pixel}
\title{The moving average dynamic of an RS index}
\usage{
dynamic_per_pixel(
  imageIndex,
  nc,
  water_scenes_only = TRUE,
  days_around_ma = 20,
  maxPixels = 1000,
  pixelFilter = NULL,
  threshold = NULL,
  maxDataPoints = 5e+06,
  returnSinglePixels = TRUE
)
}
\arguments{
\item{imageIndex}{the data list created by \code{\link[=ndi_per_image]{ndi_per_image()}}}

\item{nc}{The netcdf list returned by \code{\link[=open_netcdf]{open_netcdf()}}}

\item{water_scenes_only}{By default only pixels classified as water scene are
used for the yearly average per pixel. If False, all values are included
except for cloud and snow scenes.}

\item{days_around_ma}{The days around the moving average day (i.e. 10  means
10 days before and ten days after the actual day are used for averaging). If
NULL the number of days around the moving average is derived from the
median available values per pixel.}

\item{maxPixels}{If the number is below 1 it will be interpreted as a
proportion of all valid Pixels. Otherwise, maxPixels defines the number of
pixels used to calculate the overall dynamic.If it is smaller as the number of
valid pixels (revided by the threshold value), pixels will drawn in decreasing
order from most valid to lowest valid. If Inf, all pixels will be analysed.
MaxPixels is ignored if pixelFilter is defined.}

\item{pixelFilter}{The ID of pixels to be used (The ID of a pixel is its
its number in the the original matrix of the netcdf).}

\item{threshold}{The mininum number of valid values for one pixel to be
processed (the higher the days around moving averages are, the lower this
threshold can be)}

\item{maxDataPoints}{The number of datapoints to be part of one matrix
for moving average calculation. The default (5E+06) corresponds to 5000 pixels
in 1000 images.}

\item{returnSinglePixels}{If TRUE (default) moving average dynamic of each
pixel (up to the number of maxPixels) is returned. If False only median and
some quantiles are returned.}
}
\description{
The dynamic can include many years, all sorted by the day of the year.
The aggregation per pixel is a moving average.
}
\details{
A moving average is calculated for a day if at least two different images
are available. If the number of images is not sufficient the days_around_ma
need to be increased and the results will become smoother.
}
