% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EO_assessment_class.R
\name{EO_assessment_class}
\alias{EO_assessment_class}
\title{The assessment of eutriphication based on ranking in classes}
\usage{
EO_assessment_class(yearly_spread, nClass = 10)
}
\arguments{
\item{yearly_spread}{A dataframe of lakeName, lakeID and the yearly NDTrI
values as created by \code{\link[=ndtri_spread]{ndtri_spread()}}}

\item{nClass}{The number of class used (3 classes minimum).}
}
\value{
A list of 3:
"assessment": the yearly_spread dataframe extended by class
columns.
"breaks": The breaks used for the classification in each year.
"nClass": The total number of classes
"proportionExtreme": The proportion of lakes in the first and last class
}
\description{
The assessment of eutriphication based on ranking in classes
}
\details{
The first and last class are always defined as the least and most
eutrophicated 5\% of all lakes, respectively. The remaining breaks are equal
in range, i.e. the classes can contain different amounts of lakes.
}
