% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EO_assessment_numeric.R
\name{EO_assessment_numeric}
\alias{EO_assessment_numeric}
\title{The assessment of eutriphication based on NDTrI Values}
\usage{
EO_assessment_numeric(
  yearly_spread,
  statusYears = 3,
  shortTermYears = 3,
  longTermYears = 10
)
}
\arguments{
\item{yearly_spread}{A dataframe of lakeName, lakeID and the yearly NDTrI
values as created by \code{\link[=ndtri_spread]{ndtri_spread()}}}

\item{statusYears}{The number of years used for the moving average. By default
this is 3 years.}

\item{shortTermYears}{The number of years used for the short-term trend.}

\item{longTermYears}{The number of years used for the short-term trend.}
}
\value{
A list of 4:
"assessment": the yearly_spread dataframe  extended by status and trend
columns. For each time period, the trend column is the slope of the
regression line while error is the standard error of the slope.
"periods": number of years used for status, short-term and long-term trend
"overallMedian": median value of the whole dataset
"yearlyMedians": median values of all lakes per year
}
\description{
The assessment of eutriphication based on NDTrI Values
}
\details{
The status of NDTrI is determined as moving average over n years.

In order to rule out systematic errors based on weather conditions in a year,
different hardware or algorithms, the difference between the overall median value
of all lakes and years and the yearly median value of all lakes is added to the
NDTrI of a lake before the trends are calculated.
Subsequantially, the trend is calculated as linear regression between previous
years and adjusted NDTrI values.
For a reliable trend assessment the number of lakes needs to be large enough
(n>10).
}
