% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_trends.R
\name{find_trends}
\alias{find_trends}
\title{Identifies lake with highest trends, good or bad}
\usage{
find_trends(
  numericAssessment,
  trendType,
  decreasing = FALSE,
  outputVector = "lakeID"
)
}
\arguments{
\item{numericAssessment}{The numeric assessment list created by
\code{\link[=numericAssessment]{numericAssessment()}}}

\item{trendType}{Either "short" or "long"}

\item{decreasing}{logical, sort order increasing or decreasing (default)?}

\item{outputVector}{One of "lakeID" (default), "lakeName" or "rowNumber"}
}
\value{
Ordered vector according to the trend of lake names, lake IDs or
rownumbers
}
\description{
Identifies lake with highest trends, good or bad
}
