% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clusterLayer.R
\name{get_clusterLayer}
\alias{get_clusterLayer}
\title{Create a matrix of clusters in the same dimension as Datacube nc}
\usage{
get_clusterLayer(clusterVector, nc)
}
\arguments{
\item{clusterVector}{A named vector of clusters per pixel. The names are
are made of the index within the datacube xy-matrix combined the "p_" as
prefix (Vector is part of the returned output by \code{\link[=pixel_clusters]{pixel_clusters()}})}

\item{nc}{A netCDF list returned by \code{\link[=open_netcdf]{open_netcdf()}}. The SCL band is loaded
from this object for scene filtering.}
}
\description{
Create a matrix of clusters in the same dimension as Datacube nc
}
