% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pixel_NDTrI.R
\name{get_pixel_NDTrI}
\alias{get_pixel_NDTrI}
\title{Calculate the Normalized difference trophic index for a year}
\usage{
get_pixel_NDTrI(
  nc,
  ncImage,
  years,
  seasonMonths = 4:10,
  water_scenes_only = TRUE,
  pixelQualityThreshold = 0.8,
  lakeInfo = c("", "")
)
}
\arguments{
\item{nc}{The netCDF data list created by \code{\link[=load_netcdf]{load_netcdf()}}}

\item{ncImage}{the data list created by \code{\link[=data_per_image]{data_per_image()}}}

\item{years}{Numeric vector of all the years to be assed}

\item{seasonMonths}{Numeric vector including all months of the growing season}

\item{water_scenes_only}{By default only pixels classified as water scene are
used for the yearly average per pixel. If False, all values are included no
matter what scene they are}

\item{pixelQualityThreshold}{The quality is defined as the minimum proportion
a pixel needs to be defined as water scene to be included.}

\item{lakeInfo}{Character vector of length 2 specifying the Name and the ID
of the lake. This is not needed for any calculation but is important to
identify the lake.}
}
\description{
Calculate the Normalized difference trophic index for a year
}
