% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_per_image.R
\name{index_per_image}
\alias{index_per_image}
\title{data cube data is sliced according to timesteps, preprocessed and used to
calculate the NDTrI}
\usage{
index_per_image(
  nc,
  vars,
  indexType = "normalized_difference",
  noCloudSnow = TRUE
)
}
\arguments{
\item{nc}{The netCDF data list created by \code{\link[=load_netcdf]{load_netcdf()}}}

\item{vars}{A vector of variables to be part of the index.}

\item{indexType}{One of "normalized_difference", "quotient", "sum" or
"product" (see details)}

\item{noCloudSnow}{Logical, if the index should be set NA for pixels that
are classified as cloud or snow (SCL = c(2,3, 7:11))}
}
\description{
data cube data is sliced according to timesteps, preprocessed and used to
calculate the NDTrI
}
\details{
Sometime the reflactance bands show negative values. Those values need to be
set to 0 before the index is calculated the same is done in the sentinel hub
to harmonize the data: https://docs.sentinel-hub.com/api/latest/data/sentinel-2-l2a/

The SCL band on the other hand sometimes contains NA values which are
problematic for further processing. NA values are substituted by 0 which
is no ID for a scene

indexTypes
"normalized_difference": (var1 - var2) / (var1 + var2)
"quotient": var1 / var2
"sum": var1 + var2 + var3 + ...
"product": var1 x var2 x var3 x ...
}
