% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDTrI.R
\name{index_wise_average}
\alias{index_wise_average}
\title{NDTri per pixel}
\usage{
index_wise_average(list_of_mats, na.rm = FALSE)
}
\arguments{
\item{list_of_mats}{A list of numeric matrices of the same size}

\item{na.rm}{Logical}
}
\description{
NDTri per pixel
}
