% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_BandLayer.R
\name{load_BandLayer}
\alias{load_BandLayer}
\title{Load values of one Band from netCDF into R environment}
\usage{
load_BandLayer(nc, band, monthFilter = NULL, yearFilter = NULL)
}
\arguments{
\item{nc}{The output of \code{\link[=open_netcdf]{open_netcdf()}} function or an extension of it}

\item{band}{Character of the sentinel-2 band name}

\item{monthFilter}{Numeric vector of months to load}

\item{yearFilter}{Numeric vector of years to load}
}
\description{
Load values of one Band from netCDF into R environment
}
\details{
Sometime the reflactance bands show negative values. Those values need to be
set to 0 before the index is calculated the same is done in the sentinel hub
to harmonize the data: https://docs.sentinel-hub.com/api/latest/data/sentinel-2-l2a/
}
