% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_load_database.R
\name{load_database}
\alias{load_database}
\title{Load a database table}
\usage{
load_database(RSPath, dbName)
}
\arguments{
\item{RSPath}{Path of remoate sensing input and output folders}

\item{dbName}{Character string}
}
\value{
Function reads a csv file from RSPath/_database folder
}
\description{
Load a database table
}
