% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_netcdf.R
\name{load_netcdf}
\alias{load_netcdf}
\title{Function loads a netCDF file and imports all the important information for NDTrI}
\usage{
load_netcdf(title, path, fileName = "openEO.nc", loadAllVars = TRUE)
}
\arguments{
\item{title}{Character string. The same title that has been used to start the
job by function \code{\link[=start_openEO_job]{start_openEO_job()}}.}

\item{path}{the path of the netCDF file}

\item{fileName}{The file name including .nc ending. By default openEO
provides a file that is named "openEO.nc"}

\item{loadAllVars}{If TRUE (default) all available variable layers will be
loaded into the output list. Otherwise only bands relevant for the NDTrI will
be loaded.}
}
\value{
Returns a list of spatial and temporal variables and all the other layers of
the data cube.
}
\description{
Function loads a netCDF file and imports all the important information for NDTrI
}
