% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_per_pixel.R
\name{moving_average}
\alias{moving_average}
\title{Moving average for a sorted pixel vector with according day of the year time
vector}
\usage{
moving_average(ts_pixel, t_doy, days_around_ma)
}
\arguments{
\item{ts_pixel}{Data for one pixel of multiple images (ts)}

\item{t_doy}{Numeric vector of the day of the year according to ts_pixel
(same length and order)}

\item{days_around_ma}{The days around the moving average day (i.e. 10  means
10 days before and ten days after the actual day are used for averaging)}
}
\description{
Moving average for a sorted pixel vector with according day of the year time
vector
}
