% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_time_filter.R
\name{nc_time_filter}
\alias{nc_time_filter}
\title{Filter net cdf input by date}
\usage{
nc_time_filter(nc, tBeg, tEnd)
}
\arguments{
\item{nc}{The output of \code{\link[=load_netcdf]{load_netcdf()}} function}

\item{tBeg, tEnd}{Dates used for filtering in the format "YYYY-mm-dd"}
}
\value{
the same list as nc filtered for the defined time period
}
\description{
Filter net cdf input by date
}
