% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndi_per_image.R
\name{ndi_per_image}
\alias{ndi_per_image}
\title{Normalized Difference Index per Pixel and Image}
\usage{
ndi_per_image(nc, year, bandNames = c("B05", "B02"), monthFilter = NULL)
}
\arguments{
\item{nc}{The netCDF data list created by \code{\link[=open_netcdf]{open_netcdf()}}}

\item{year}{Numeric value of year to be assessed.}

\item{bandNames}{A character vector of length 2 defining the sentinel-2
bands. To see what bands are available in the datacube run 'nc$bands'.
The default is "B05" and "B02" which is used for NDTrI calculation.}

\item{monthFilter}{Optional numeric vector of months used for the index.
If NULL all months are used.}
}
\description{
Normalized Difference Index per Pixel and Image
}
\details{
The Index will be calculated by
NDI = (bandNames_1 - bandNames_2) / (bandNames_1 + bandNames_2)

Negative reflactance values are set to zero. See details of
\code{\link[=load_BandLayer]{load_BandLayer()}}

The SCL band on the other hand sometimes contains NA values which are
problematic for further processing. NA values are substituted by 0 which
is no ID for a scene
}
