% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDTrI.R
\name{NDTrI}
\alias{NDTrI}
\title{Calculate the Normalized difference trophic index for a year}
\usage{
NDTrI(
  nc,
  ncImage,
  waterLayer,
  year,
  seasonMonths = c(4, 10),
  water_scenes_only = TRUE,
  pixelQualityThreshold = 0.8
)
}
\arguments{
\item{nc}{The netCDF data list created by \code{\link[=load_netcdf]{load_netcdf()}}}

\item{ncImage}{the data list created by \code{\link[=data_per_image]{data_per_image()}}}

\item{waterLayer}{The water scene proportion layer as part of the SCL list
created by \code{\link[=waterscene_proportion]{waterscene_proportion()}}}

\item{year}{Numeric value of the year}

\item{seasonMonths}{Numeric value of length two defining the first and last
month (both included) of the growing season.}

\item{water_scenes_only}{By default only pixels classified as water scene are
used for the yearly average per pixel. If False, all values are included no
matter what scene they are}

\item{pixelQualityThreshold}{The quality is defined as the minimum proportion
a pixel needs to be defined as water scene to be included.}
}
\description{
Calculate the Normalized difference trophic index for a year
}
