% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_per_image.R
\name{ndtri_per_image}
\alias{ndtri_per_image}
\title{data cube data is sliced according to timesteps, preprocessed and used to
calculate the NDTrI}
\usage{
ndtri_per_image(nc)
}
\arguments{
\item{nc}{The netCDF data list created by \code{\link[=load_netcdf]{load_netcdf()}}}
}
\description{
data cube data is sliced according to timesteps, preprocessed and used to
calculate the NDTrI
}
\details{
Sometime the reflactance bands show negative values. Those values need to be
set to 0 before the index is calculated the same is done in the sentinel hub
to harmonize the data: https://docs.sentinel-hub.com/api/latest/data/sentinel-2-l2a/

The SCL band on the other hand sometimes contains NA values which are
problematic for further processing. NA values are substituted by 0 which
is no ID for a scene
}
