% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndtri_spread.R
\name{ndtri_spread}
\alias{ndtri_spread}
\title{Spread the NDTrI database by year}
\usage{
ndtri_spread(
  df,
  aggregationType = "modus",
  filterPixels = 0,
  filterImages = 0,
  filterQuality = 0
)
}
\arguments{
\item{df}{The database dataframe}

\item{aggregationType}{Either "modus" (default) or "median"}

\item{filterPixels}{Integer: Filter for lake trophic state assessment based
on an equal or greater number of "water" pixels per lake. Water pixels are
pixels that fullfill the quality threshold value.}

\item{filterImages}{Integer: Filter for lake trophic state assessment based on an
equal or greater number of images per season}

\item{filterQuality}{Number between 0 and 1: Filter for lake trophic state
assessment based on an equal or greater quality threshold (proportion of
pixels that are identified as water by the SCL band)}
}
\value{
A data frame with columns lakeName, lakeID and one column per year
that lists the NDTrI of the lake --> One row per lake
}
\description{
Spread the NDTrI database by year
}
