% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_properties_names.R
\docType{data}
\name{NDTrIColors}
\alias{NDTrIColors}
\title{Colors for the Index}
\format{
A dataframe of two columns. "ndtri" listing the values of the
noramlized diference trophic index and "color" specifying the corresponding
colors
}
\usage{
NDTrIColors
}
\description{
Colors for the Index
}
\keyword{datasets}
