% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndtriLake_to_df.R
\name{ndtriLake_to_df}
\alias{ndtriLake_to_df}
\title{Reshapes NDTrI Lake list to dataframe}
\usage{
ndtriLake_to_df(ndtriLake)
}
\arguments{
\item{ndtriLake}{A list of NDTrI per pixel of one or more years, created by
\code{\link[=aggregate_NDTrI]{aggregate_NDTrI()}}}
}
\description{
Reshapes NDTrI Lake list to dataframe
}
