% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneLake_df.R
\name{oneLake_df}
\alias{oneLake_df}
\title{Reshapes NDTrI Lake list to dataframe}
\usage{
oneLake_df(lakeIndexList)
}
\arguments{
\item{lakeIndexList}{A list of yearly lake index lists of one lake created by
\code{\link[=seasonal_index_per_lake]{seasonal_index_per_lake()}}.}
}
\description{
Reshapes NDTrI Lake list to dataframe
}
