% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_netcdf.R
\name{open_netcdf}
\alias{open_netcdf}
\title{Open netCDF file}
\usage{
open_netcdf(filePath, fileName = "openEO.nc")
}
\arguments{
\item{filePath}{the path of the netCDF file}

\item{fileName}{The file name including .nc ending. By default openEO
provides a file that is named "openEO.nc"}
}
\value{
Returns a list of spatial and temporal variables, the netcdf CRS (coordinate
reference system) and the spatRaster of all layers
}
\description{
Open netCDF file
}
