% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_class_assessment.R
\name{plot_class_assessment}
\alias{plot_class_assessment}
\title{Plot class assessment results of lakes}
\usage{
plot_class_assessment(
  class_assessment,
  lakeNames = NULL,
  lakeIDs = NULL,
  rowNumbers = NULL
)
}
\arguments{
\item{class_assessment}{The numeric assessment list created by
\code{\link[=discreteClassAssessment]{discreteClassAssessment()}}}

\item{lakeNames, lakeIDs, rowNumbers}{One or more of those can be defined as
character vector for lakeNames and lakeIDs or numeric vectors for rowNUmbers.
Different filters can be combined}
}
\description{
Lake can be selected either by Name, ID or the rownumber in the assessment table
}
