% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_image_dates.R
\name{plot_image_dates}
\alias{plot_image_dates}
\title{Plot of images per day and year}
\usage{
plot_image_dates(vDate, vCoverage = NULL)
}
\arguments{
\item{vDate}{A vector of image dates}

\item{vCoverage}{A dataframe of scene proportion per image created by
\code{\link[=nc_scene_per_image]{nc_scene_per_image()}}}
}
\description{
Plot of images per day and year
}
