% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lake_ndtri_histogram.R
\name{plot_lake_ndtri_histogram}
\alias{plot_lake_ndtri_histogram}
\title{Plot the histrom of all pixel NDTrIs of a lake for one year}
\usage{
plot_lake_ndtri_histogram(ndtriLake, year)
}
\arguments{
\item{ndtriLake}{the output of \code{\link[=aggregate_NDTrI]{aggregate_NDTrI()}}}

\item{year}{The year to be plotted as numeric}
}
\description{
Plot the histrom of all pixel NDTrIs of a lake for one year
}
