% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_leaflet.R
\name{save_leaflet}
\alias{save_leaflet}
\title{Save a leaflet map as a self-contained HTML file}
\usage{
save_leaflet(x, file)
}
\arguments{
\item{x}{A \code{leaflet}  object.}

\item{file}{Character scalar. Output HTML file including path.}
}
\value{
The return value of \code{\link[htmlwidgets:saveWidget]{htmlwidgets::saveWidget()}}. The main side effect
is writing \code{file}.
}
\description{
Adds a mobile-friendly viewport meta tag to a leaflet widget and saves it as a
self-contained HTML file.
}
