% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_properties_names.R
\docType{data}
\name{sceneIDs}
\alias{sceneIDs}
\title{Sentinel-2 IDs of the SCL Band}
\format{
A data frame of two columsn 1) SCL ID and 2) SCL name
}
\usage{
sceneIDs
}
\description{
Sentinel-2 IDs of the SCL Band
}
\keyword{datasets}
