% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_per_pixel.R
\name{scl_filter}
\alias{scl_filter}
\title{Set values NA by SCL}
\usage{
scl_filter(indexImage, sclImage, bands, invert = FALSE)
}
\arguments{
\item{indexImage, sclImage}{Index and SCL layer of the image}

\item{bands}{Numeric vector specifiyng the SCL categories to filtered for}

\item{invert}{If TRUE the filtering is inverted -> specefied bands are removed}
}
\description{
Set values NA by SCL
}
