% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scl_mask.R
\name{scl_mask}
\alias{scl_mask}
\title{Mask index values by Sentinel-2 SCL classes}
\usage{
scl_mask(indexImage, sclImage, sclCategories, invert = FALSE)
}
\arguments{
\item{indexImage}{Numeric matrix of index values.}

\item{sclImage}{Matrix of SCL values with the same dimensions as \code{indexImage}.}

\item{sclCategories}{Numeric vector of SCL classes to keep in the current
implementation.}

\item{invert}{If TRUE the filtering is inverted. All categories except the
specified ones are removed}
}
\description{
Sets index pixels to \code{NA} based on matching Sentinel-2 scene-classification
values.
}
\keyword{internal}
