% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_index_per_lake.R
\name{seasonal_index_per_lake}
\alias{seasonal_index_per_lake}
\title{Aggregation of all pixel NDTrIs into one lake NDTrI}
\usage{
seasonal_index_per_lake(seasonIndex, lakeName = NULL, lakeID = NULL)
}
\arguments{
\item{seasonIndex}{A list of NDTrI per pixel of one or more years, created by
\code{\link[=seasonal_index_per_pixel]{seasonal_index_per_pixel()}}}

\item{lakeName}{Character string specifying the name of the lake.
Defaults to \code{NULL}, which assigns \code{NA}.}

\item{lakeID}{Character string specifying a unique identifier for the lake.
Defaults to \code{NULL}, which triggers automatic generation in the format
\code{"<mean_x>_<mean_y>"} coordinate.}
}
\description{
Aggregation of all pixel NDTrIs into one lake NDTrI
}
