% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_index_per_pixel.R
\name{seasonal_index_per_pixel}
\alias{seasonal_index_per_pixel}
\title{Calculate an index per pixel averaged over time}
\usage{
seasonal_index_per_pixel(
  imageIndex,
  nc,
  water_scenes_only = TRUE,
  pixelQualityThreshold = 0.8
)
}
\arguments{
\item{imageIndex}{the data list created by \code{\link[=ndi_per_image]{ndi_per_image()}}}

\item{nc}{The netcdf list returned by \code{\link[=open_netcdf]{open_netcdf()}}}

\item{water_scenes_only}{By default only pixels classified as water scene are
used for the yearly average per pixel. If False, all values are included no
except for cloud and snow scenes.}

\item{pixelQualityThreshold}{The quality is defined as the minimum proportion
a pixel needs to be defined as water scene to be included.}
}
\value{
A list of
\itemize{
\item lakeInfo: A character of lake Name and ID or whatever is provided as
meta data#
\item A list per season
\itemize{
\item RSindex: A Matrix of the whole rechtangle containing the index per pixel
for the whole season
\item QualityThreshold: The minimum proportion a pixel is required to be
identified as water during one season
\item minValuesPerPixel: The mininum number of images used for index calculation
ofone pixel (Quality_threshold x Images_of_Season)
\item validPixel: The number of pixels above Quality Threshold
}
}
}
\description{
Calculate an index per pixel averaged over time
}
