% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_properties_names.R
\docType{data}
\name{tenClassColors}
\alias{tenClassColors}
\title{Colors for the Classes}
\format{
A data frame of two columns. "class" listing the classes 1 to 10 and
"color" specifying the corresponding colors
}
\usage{
tenClassColors
}
\description{
Colors for the Classes
}
\keyword{datasets}
