% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_properties_names.R
\docType{data}
\name{tenClusterColors}
\alias{tenClusterColors}
\title{Colors for the Clusters}
\format{
A data frame of two columns. "cluster" listing the clusters 1 to 10
and "color" specifying the corresponding colors
}
\usage{
tenClusterColors
}
\description{
Colors for the Clusters
}
\keyword{datasets}
