% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_from.R
\name{bbox_from_point}
\alias{bbox_from_point}
\title{Bounding Box from point coordinate}
\usage{
bbox_from_point(lon, lat, meters = 20)
}
\arguments{
\item{lon, lat}{Longitude and latitude in WGS 84 EPSG:4326}

\item{meters}{The distance around the point to create the bounding box.
Needs to be > 0.}
}
\description{
Bounding Box from point coordinate
}
