% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_from.R
\name{bbox_from_rectangle}
\alias{bbox_from_rectangle}
\title{Bounding Box from rectangle}
\usage{
bbox_from_rectangle(top_lat, bottom_lat, left_lon, right_lon, meters = 20)
}
\arguments{
\item{top_lat, bottom_lat}{Latitude of the bottom and top of the rectangle
in WGS 84 EPSG:4326 (--> Top value is higher than bottom value)}

\item{left_lon, right_lon}{Longitude of the left and right side of the
rectangle in WGS 84 EPSG:4326 (--> right side value is higher than left side
value)}

\item{meters}{The distance around the point to create the bounding box.
Needs to be > 0.}
}
\description{
Bounding Box from rectangle
}
