% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EO_assessment_dynamic.R
\name{best_nk}
\alias{best_nk}
\title{Find the number of clusters to use in kmeans}
\usage{
best_nk(ts_table, kMax = 10, improvement_objective = 10, plot_result = TRUE)
}
\arguments{
\item{ts_table}{A table of timeseries data per pixel created by
\code{\link[=dynamic_per_pixel]{dynamic_per_pixel()}}}

\item{kMax}{Tha maximum number of clusters to be tested}

\item{improvement_objective}{The improvement of within cluster sum of squares
(WCSS) by addition of one cluster in \%.}

\item{plot_result}{}
}
\value{
This function plots the improvement values and returns the recommended
number of cluster to use
}
\description{
Based on the reduction of within cluster sum of squares
}
