% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_database.R
\name{create_database}
\alias{create_database}
\title{Creates a new database table}
\usage{
create_database(RSPath, included_folders, dbName)
}
\arguments{
\item{RSPath}{Path of remoate sensing input and output folders}

\item{included_folders}{Subfolders or the output folder which are part of the
new databse}

\item{dbName}{Character string}
}
\value{
Function saves a csv file in the RSPath/_database folder
}
\description{
Creates a new database table
}
