% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EO_assessment_numeric.R
\name{determine_status}
\alias{determine_status}
\title{The status of NDTrI is determined as moving average over n years}
\usage{
determine_status(yearly_spread, n_years = 3)
}
\arguments{
\item{yearly_spread}{A dataframe of lakeName, lakeID and the yearly NDTrI
values as created by \code{\link[=ndtri_spread]{ndtri_spread()}}}

\item{n_years}{The number of years used for the moving average. By defaul
this is 3 years.}
}
\value{
A list of 2:

A matrix, where number of rows equal the number of lakes (in the same order
as in the yearly_spread input table) and number of correspond to the number
of years assessed
}
\description{
The status of NDTrI is determined as moving average over n years
}
