% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EO_assessment_numeric.R
\name{determine_trend}
\alias{determine_trend}
\title{The Trend of NDTrI is determined within a coelleciton of lakes in short and
long-term}
\usage{
determine_trend(yearly_spread = yearly_spread, shortTerm = 3, longTerm = 10)
}
\arguments{
\item{yearly_spread}{A dataframe of lakeName, lakeID and the yearly NDTrI
values as created by \code{\link[=ndtri_spread]{ndtri_spread()}}}

\item{shortTerm}{The number of years used for the short-term trend.}

\item{longTerm}{The number of years used for the short-term trend.}
}
\value{
A list of 4:
"trends" is a matrix with one row per lake in the as in the
yearlyspread input table and two columns for each trend, short-term and
long-term. The trend is the slope of the regression line, and error is the
standard error of the slope.
"periods": number of years used for short-term and long-term trend
"overallMedian": median value of the whole dataset
"yearlyMedians": median values of all lakes per year
}
\description{
The Trend of NDTrI is determined within a coelleciton of lakes in short and
long-term
}
\details{
In order to rule out systematic errors based on weather conditions in a year,
different hardware or algorithms, the difference between the overall median
value of all lakes and years and the yearly median value of all lakes is
added to the NDTrI of a lake before the trends are calculated. Median values
are based on lakes or pixels for which data is available in all considered
years.
Subsequantly, the trend is calculated as linear regression between previous
years and adjusted NDTrI values.
For a reliable trend assessment the number of lakes needs to be large enough
(n>10).
}
