% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_bbox.R
\name{display_geometry}
\alias{display_geometry}
\title{Display the Bounding Box}
\usage{
display_geometry(geom, zoom = 15)
}
\arguments{
\item{geom}{The geometry to display. Either a bounding box or a polygon}

\item{zoom}{The zoom level of the initial map}
}
\description{
Display the Bounding Box
}
\details{
Bounding Box as a names vector or
c("north" = lat1, "east"  = lon1, "south" = lat2, "west" = lon2) or a list of
pairs list(c(lat1, lng2), c(lat2, lon2), c(lat3, lon3))
}
