% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_openEO_job.R
\name{download_openEO_job}
\alias{download_openEO_job}
\title{Downloads an openEO job}
\usage{
download_openEO_job(
  title,
  path,
  recentJob = TRUE,
  ID = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{title}{Character string. The same title that has been used to start the
job by function \code{\link[=start_openEO_job]{start_openEO_job()}}. If ID is specified the title is still
used for the folder.}

\item{path}{The path where to download the data to}

\item{recentJob}{If TRUE, only the last 10 jobs are searched for the title}

\item{ID}{If specified the Job ID is used instead of jobs title}

\item{overwrite}{If TRUE (Default is FALSE) existing folder will be used and
previous data might be overwritten}
}
\value{
Function only returns the information to which folder data is downloaded
}
\description{
This function downloads a job from openEO if it is finished. You need to be
connected and logged in to run this function.
}
