% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_per_pixel.R
\name{dynamic_per_pixel}
\alias{dynamic_per_pixel}
\title{A band or index sorted by the day of the year}
\usage{
dynamic_per_pixel(
  ncImage,
  t_date,
  years,
  water_scenes_only = TRUE,
  days_around_ma = 20,
  nMax = 1000,
  threshold = NULL,
  lakeInfo = c("", "")
)
}
\arguments{
\item{ncImage}{the data list created by \code{\link[=ndi_per_image]{ndi_per_image()}},
which must contain the SCL layer if water_scenes_only is TRUE.}

\item{t_date}{A vector of the dates of images. This is part of the output
of \code{\link[=load_netcdf]{load_netcdf()}}, with the list name "t_date".}

\item{years}{Numeric vector of all the years to be included}

\item{water_scenes_only}{By default only pixels classified as water scene are
used for the yearly average per pixel. If False, all values are included
except for cloud and snow scenes.}

\item{days_around_ma}{The days around the moving average day (i.e. 10  means
10 days before and ten days after the actual day are used for averaging)}

\item{nMax}{Maximum number of pixels (randomly chosen from available pixels).
If Inf, all pixels will be analysed.}

\item{threshold}{The mininum number of valid values for one pixel to be
processed (the higher the days around moving averages are, the lower this
threshold can be)}

\item{lakeInfo}{Character vector of length 2 specifying the Name and the ID
of the lake. This is not needed for any calculation but is important to
identify the lake.}
}
\description{
The dynamic can include many years, all sorted by the day of the year.
The aggregation per pixel is a moving average.
}
