% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_per_pixel.R
\name{images_per_ma}
\alias{images_per_ma}
\title{Subset of all images to be used for moving averages for each day}
\usage{
images_per_ma(t_doy, days_around_ma)
}
\arguments{
\item{t_doy}{Numeric vector of the day of the year according to ts_pixel
(same length and order)}

\item{days_around_ma}{The days around the moving average day (i.e. 10  means
10 days before and ten days after the actual day are used for averaging)}
}
\description{
Subset of all images to be used for moving averages for each day
}
