% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_per_image.R
\name{load_bandLayer}
\alias{load_bandLayer}
\title{Load one layer of a band from a datacrube}
\usage{
load_bandLayer(nc, bandName, timeStep)
}
\arguments{
\item{nc}{The netCDF data list created by \code{\link[=load_netcdf]{load_netcdf()}}}

\item{bandName}{A character vector of length 2 defining the sentinel-2
bands. To see what bands are available in the datacube run 'names(nc)'.}

\item{timeStep}{Integer of the layer (timestep in a timeseries)}
}
\description{
Load one layer of a band from a datacrube
}
\details{
Sometime the reflactance bands show negative values. Those values need to be
set to 0 before the index is calculated the same is done in the sentinel hub
to harmonize the data: https://docs.sentinel-hub.com/api/latest/data/sentinel-2-l2a/
}
