% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_layer_map_layer.R
\name{map_layer}
\alias{map_layer}
\title{Leaflet Map one layer (Matrix) described by a netCDF}
\usage{
map_layer(
  ncLayer,
  nc,
  aboveValues = NULL,
  highestValue = NULL,
  aboveColors = NULL,
  valueRange = NULL,
  legendTitle = NULL,
  plotLegend = TRUE
)
}
\arguments{
\item{ncLayer}{This is a layer that corresponds to the x and y dimensions of
the netCDF file. It can be a band of netCDF or further processed layer
created by \code{\link[=waterscene_proportion]{waterscene_proportion()}}}

\item{nc}{The netCDF data list created by \code{\link[=load_netcdf]{load_netcdf()}}}

\item{aboveValues}{Lower limits of value classes. If NULL (default)
numerical values will be used.}

\item{highestValue}{If Null the highest value is derived by the maximum of
all available layer values. It can also be specified manually.}

\item{aboveColors}{A vector of R-colors names corresponding to the value
classes.}

\item{valueRange}{Minimum and maximum values used for color scale of numeric
values.}

\item{legendTitle}{Character string for legend title}

\item{plotLegend}{If TRUE legend will be plotted. If now aboveValues and no
valueRange are spefecied. No legend will be plotted.}
}
\description{
The plot is the layer on top of a leaflet map. The color scheme can
be done for numeric or categorical values.
}
\details{
Use case: Proportions of a scene.
The above values and correspoding colors need to be specified. If for example
values between 0 and 0.5 (50\%) should bewhite, values between 0.5 and 0.75
should be yellow and values above 0.75 should be purple, aboveValues are
c(0, 0.5, 0.75) and the aboveColors arec("white", "yellow", "purple").
In that case a legendTitle is recommended to explain what is displayed.
}
