% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_scene_per_image.R
\name{nc_scene_per_image}
\alias{nc_scene_per_image}
\title{Calculates the proportion of a specified scene in an image}
\usage{
nc_scene_per_image(nc, scene = "clouds")
}
\arguments{
\item{nc}{The output of \code{\link[=load_netcdf]{load_netcdf()}} function}

\item{scene}{Character defining the scene. Can be one of "clouds", "water",
"snow" or "shadows"}
}
\description{
Calculates the proportion of a specified scene in an image
}
