% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_per_image.R
\name{ndi_per_image}
\alias{ndi_per_image}
\title{data cube data is sliced according to timesteps, preprocessed and used to
calculate a normalized difference index}
\usage{
ndi_per_image(nc, bandNames = c("B05", "B02"), keep_bands = FALSE)
}
\arguments{
\item{nc}{The netCDF data list created by \code{\link[=load_netcdf]{load_netcdf()}}}

\item{bandNames}{A character vector of length 2 defining the sentinel-2
bands. To see what bands are available in the datacube run 'names(nc)'.
The default is "B05" and "B02" which is used for NDTrI calculation.}

\item{keep_bands}{If True the band 2, 5 and SCL data is part of the list (as
matrix per image)}
}
\description{
data cube data is sliced according to timesteps, preprocessed and used to
calculate a normalized difference index
}
\details{
The Index will be calculated by
NDI = (bandNames_1 - bandNames_2) / (bandNames_1 + bandNames_2)

Negative reflactance values are set to zero. See details of
\code{\link[=load_bandLayer]{load_bandLayer()}}

The SCL band on the other hand sometimes contains NA values which are
problematic for further processing. NA values are substituted by 0 which
is no ID for a scene
}
