% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_analysis.R
\name{pixel_clusters}
\alias{pixel_clusters}
\title{cluster analysis (k-means) and map layer}
\usage{
pixel_clusters(
  pixelDynamic,
  nc,
  k,
  iter.max = 20,
  nstart = 10,
  correlate_first = FALSE,
  whole_dynamic = FALSE
)
}
\arguments{
\item{pixelDynamic}{The pixel data table of the output of \code{\link[=dynamic_per_pixel]{dynamic_per_pixel()}}}

\item{nc}{The netCDF data list created by \code{\link[=open_netcdf]{open_netcdf()}}}

\item{k}{The number of Clusters}

\item{iter.max, nstart}{Arguments of \code{\link[=kmeans]{kmeans()}}}

\item{correlate_first}{If TRUE, Pearson correlation between pixels will be
used for clustering instead of Index values. This results in a clustering
based on different shapes of the dynamic.}

\item{whole_dynamic}{If TRUE all 365 days average values will be used to
compare pixels. Otherwise the number of days will be reduced if the number
of pixels is too high (> 100 000).}
}
\value{
A list of the k-means output and a layer of clustered pixels
}
\description{
cluster analysis (k-means) and map layer
}
