% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EO_assessment_dynamic.R
\name{pixel_clusters}
\alias{pixel_clusters}
\title{cluster analysis (k-means) and map layer}
\usage{
pixel_clusters(dpp, nc, k, iter.max = 20, nstart = 10)
}
\arguments{
\item{dpp}{The output of \code{\link[=dynamic_per_pixel]{dynamic_per_pixel()}}}

\item{nc}{The netCDF data list created by \code{\link[=load_netcdf]{load_netcdf()}}}

\item{k}{The number of Clusters}

\item{iter.max, nstart}{Arguments of \code{\link[=kmeans]{kmeans()}}}
}
\value{
A list of the k-means output and a layer of clusted pixels
}
\description{
cluster analysis (k-means) and map layer
}
