% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dynamic.R
\name{plot_dynamic}
\alias{plot_dynamic}
\title{Plot the moving average dynamic of an index}
\usage{
plot_dynamic(dpp, lakeName = "", ylab = "Index", pixelClusters = NULL)
}
\arguments{
\item{dpp}{The dynamic per pixel output by \code{\link[=dynamic_per_pixel]{dynamic_per_pixel()}}}

\item{lakeName}{String defining the title of the plot}

\item{ylab}{y axis name}

\item{pixelClusters}{Optional: Named Integer vector of clusters. Names
according to the pixel names (column names in the dpp$moving_average table).
If pixelClusters are provided the dynamic is plotted for each cluster
separately.}
}
\description{
Plot the moving average dynamic of an index
}
\details{
A maximum of 5000 timeseries will be drawn as single lines. If more data is
available 5000 timeseries are randomly drawn. The transparency of each single line
depends on the number of lines to be drawn (the more lines, the higher the
transparency)
}
