% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lake_index_histogram.R
\name{plot_lake_index_histogram}
\alias{plot_lake_index_histogram}
\title{Plot the histrom of all pixel NDTrIs of a lake for one year}
\usage{
plot_lake_index_histogram(lakeIndex, lakeName, indexName)
}
\arguments{
\item{lakeIndex}{the output of \code{\link[=seasonal_index_per_lake]{seasonal_index_per_lake()}}}

\item{lakeName, indexName}{Character strings of the lake and index name used
in the plot}
}
\description{
Plot the histrom of all pixel NDTrIs of a lake for one year
}
