% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_numeric_assessment.R
\name{plot_numeric_assessment}
\alias{plot_numeric_assessment}
\title{Plot the numeric EO eutriphication assessment for one lake}
\usage{
plot_numeric_assessment(
  numericData,
  lakeName = NULL,
  lakeID = NULL,
  rowNumber = NULL,
  ylab = "Index"
)
}
\arguments{
\item{numericData}{The numeric assessment list created by
\code{\link[=numericAssessment]{numericAssessment()}}}

\item{lakeName, lakeID, rowNumber}{One of those need to be defined to select
the lake to be plotted}

\item{ylab}{Y axis lable}
}
\description{
Lake can be selected either by Name, ID or the rownumber in the assessment table
}
