% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_numeric_assessment.R
\name{plot_numeric_assessment}
\alias{plot_numeric_assessment}
\title{Plot the numeric EO eutriphication assessment for one lake}
\usage{
plot_numeric_assessment(
  numeric_assessment,
  lakeName = NULL,
  lakeID = NULL,
  rowNumber = NULL
)
}
\arguments{
\item{numeric_assessment}{The numeric assessment list created by
\link{EO_assessment_numeric}}

\item{lakeName, lakeID, rowNumber}{One of those need to be defined to select
the lake to be plotted}
}
\description{
Lake can be selected either by Name, ID or the rownumber in the assessment table
}
