% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_layer_map_layer.R
\name{plot_scale}
\alias{plot_scale}
\title{Define plot region scale}
\usage{
plot_scale(pin_w, pin_h)
}
\arguments{
\item{pin_w}{widthsize}

\item{pin_h}{hightsize}
}
\description{
Define plot region scale
}
\details{
both values will be treated relatively and will adjust the plot margin to
the device size
}
