% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_analysis.R
\name{prepare_for_clustering}
\alias{prepare_for_clustering}
\title{Pre-processing done before cluster analysis}
\usage{
prepare_for_clustering(
  moving_averages_matrix,
  correlate_first = FALSE,
  maxPixels = NULL
)
}
\arguments{
\item{moving_averages_matrix}{A matrix of p pixels as column and 365 days as
rows of the moving averages of one index as created by \code{\link[=dynamic_per_pixel]{dynamic_per_pixel()}}}

\item{correlate_first}{If TRUE, Pearson correlation between pixels will be
used for clustering instead of Index values. This results in a clustering
based on different shapes of the dynamic.}

\item{maxPixels}{The maximum numbers of pixel used for cluster analysis.
Leave NULL to use all pixels.}
}
\value{
A matrix either with data per pixels or correlation between pixels
used for following cluster analysis
}
\description{
Pre-processing done before cluster analysis
}
