% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_class_assessment.R
\name{rescale_classColors}
\alias{rescale_classColors}
\title{Rescale class colors}
\usage{
rescale_classColors(nClass)
}
\arguments{
\item{nClass}{Number of Classes}
}
\description{
Rescale class colors
}
