% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dynamic.R
\name{rm_na_for_polygon}
\alias{rm_na_for_polygon}
\title{Remove NA values for polygon plotting}
\usage{
rm_na_for_polygon(x_v, y_v)
}
\arguments{
\item{x_v, y_v}{vectors of x and y values of the polygon}
}
\description{
Remove NA values for polygon plotting
}
