% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_lake_data.R
\name{save_lake_data}
\alias{save_lake_data}
\title{Save or update data of the lake}
\usage{
save_lake_data(
  RSPath,
  nc,
  ncImage,
  sceneProportion,
  ndtriPixels,
  ndtriLake,
  save_plots = TRUE
)
}
\arguments{
\item{RSPath}{The path of the Remote Sensing data. One level above input
and output folder}

\item{nc}{The netCDF data list created by \code{\link[=load_netcdf]{load_netcdf()}}}

\item{ncImage}{the data list created by \code{\link[=data_per_image]{data_per_image()}}}

\item{sceneProportion}{A list of the proportions of different scene, created
by \code{\link[=waterscene_proportion]{waterscene_proportion()}}}

\item{ndtriPixels}{A list of NDTrI per pixel of one or more years, created by
\code{\link[=get_pixel_NDTrI]{get_pixel_NDTrI()}}}

\item{ndtriLake}{A list of NDTrI data per lake of one or more years, created by
\code{\link[=aggregate_NDTrI]{aggregate_NDTrI()}}}

\item{save_plots}{If TRUE all maps and plots are saved.}
}
\description{
Maps are saved as .html file, plots as .png, R objects as .rds and a
summary is saved as csv table. The folder where data is saved is a subfolder
of RSPath/output and is named by a combination of lake name and lake ID.
}
