% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_leaflet.R
\name{save_leaflet}
\alias{save_leaflet}
\title{Save a leaflet as html}
\usage{
save_leaflet(x, file)
}
\arguments{
\item{x}{Leaflet object to be saved}

\item{file}{Filename including path}
}
\description{
By adding content to the html it is assured that icons are scaled according
to the device. In this way the map can be used by mobile devices.
}
