% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_index_per_pixel.R
\name{seasonal_index_per_pixel}
\alias{seasonal_index_per_pixel}
\title{Calculate an index per pixel averaged over time}
\usage{
seasonal_index_per_pixel(
  nc,
  ncImage,
  years,
  seasonMonths = 4:10,
  water_scenes_only = TRUE,
  pixelQualityThreshold = 0.8,
  lakeInfo = c("", "")
)
}
\arguments{
\item{nc}{The netCDF data list created by \code{\link[=load_netcdf]{load_netcdf()}}}

\item{ncImage}{the data list created by \code{\link[=ndi_per_image]{ndi_per_image()}}}

\item{years}{Numeric vector of all the years to be assessed}

\item{seasonMonths}{Numeric vector including all months of the growing season}

\item{water_scenes_only}{By default only pixels classified as water scene are
used for the yearly average per pixel. If False, all values are included no
except for cloud and snow scenes.}

\item{pixelQualityThreshold}{The quality is defined as the minimum proportion
a pixel needs to be defined as water scene to be included.}

\item{lakeInfo}{Character vector of length 2 specifying the Name and the ID
of the lake. This is not needed for any calculation but is important to
identify the lake.}
}
\description{
Calculate an index per pixel averaged over time
}
