% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_openEO_job.R
\name{start_openEO_job}
\alias{start_openEO_job}
\title{Start Job on openEO plattform}
\usage{
start_openEO_job(
  title,
  geom,
  tBeg,
  tEnd,
  polygon,
  crs = 4326,
  collection = "SENTINEL2_L2A",
  bands = list("B02", "B05", "SCL"),
  relativeOrbitNumber = NULL,
  tileID = NULL,
  outputFormat = "netCDF",
  uniqueTitle = TRUE
)
}
\arguments{
\item{title}{A character defining the job title (can be a lake name, id). It
is advised to use a unique name (i.e. including a timestamp) that is used on
openEO and as a folder on the local machine}

\item{geom}{Numerical vector defining the bounding box in the coordinate
reference system as defined in crs. The order of the values is north, east,
south, west. Alternatively, geom can be a list of 3 or more
latitude-longitude pairs, each pair is a vector of length 2 in the list to
create a polyong (which is actually not bounding box any longer...)}

\item{tBeg, tEnd}{Dates in the format "YYYY-MM-DD" which define the temporal
extent of the data to be downloaded. Left side (tBeg) is included, right side
(tEnd) is excluded from range.}

\item{crs}{A numerical value defining the coordinate reference system. By
default this is 4326 (WGS 84). See here for further information:
https://epsg.io/4326}

\item{collection}{One of the available collections on openEO
(SENTINEL2_L2A is default). The collections can be listed by
\code{\link[openeo:list_collections]{openeo::list_collections()}}}

\item{bands}{A list of characters defining The bands to be downloaded.
For the normalized difference trophic index bands 2, 5 and the scene
classification (SCL) are required. Set to NULL to download all bands.}

\item{outputFormat}{On of the available output formats (netcdf is default).
The formats can be listed by \code{\link[openeo:list_file_formats]{openeo::list_file_formats()}}. Other alternatives
for higher dimensional data cubes are "GeoJSON" or "GTiff".}

\item{uniqueTitle}{If TRUE, the previous jobs are checked. If the title has
already been used, it is suffixed by an underscore and the lowest possible
number}
}
\value{
This function only returns the title of the job that is started on openEO.
To download the job use function \link{download_openEO_job}
}
\description{
This function will implement and start a job on openEO. You need to be
connected and logged in to run this function.
}
\details{
Laitude-Longitude Pairs are switched to longitude-latitude which is the
required order of openEO, however providing the data as latitude-longitude
is more convient beacuse it is the google maps output format
}
