% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterscene_proportion.R
\name{waterscene_proportion}
\alias{waterscene_proportion}
\title{Proprtions of all scenes and dervied water scene proportion}
\usage{
waterscene_proportion(scl_image)
}
\arguments{
\item{scl_image}{The SCL layer of the imagewise data, created by
\code{\link[=data_per_image]{data_per_image()}}}
}
\value{
A list of 3:
\itemize{
\item the water scene proportion per pixel
\item the disturbance scene proportion per pixel
\item A list of all individual scene proportions per pixel
}
}
\description{
Proprtions of all scenes and dervied water scene proportion
}
\details{
The water proportion is defined as all pixels that are "correctly" classified
as water. Thus, disturbances that occur but are not a false classification
like clouds, snow or topographic impacts need to be removed before. After
removing those pixels, the water classifiation should be 100\% for a perfect
pixel
}
